requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
python-docx
openpyxl
matplotlib
reportlab
Pillow
